+++ src/biodwh2-core/src/test/java/de/unibi/agbi/biodwh2/core/WorkspaceTest.java
@@ -3,7 +3,6 @@ package de.unibi.agbi.biodwh2.core;
 import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -15,7 +14,6 @@ import java.text.DecimalFormat;
 import java.util.concurrent.ForkJoinPool;
 import java.util.stream.Stream;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 @Disabled("For development and demonstration purposes only")
@@ -24,7 +22,6 @@ class WorkspaceTest {
 
     private Workspace workspace;
     private final String[] ids = {"Mock1", "Mock2", "Mock3"};
-
     private DecimalFormat df;
 
     @BeforeAll
@@ -39,20 +36,15 @@ class WorkspaceTest {
     void processDataSourcesInParallel(boolean isParallel, int numIterations) {
 
         long start, stop, elapsed, averageExecutionTime;
-
         if (numIterations > 0) {
             numIterations = numIterations;
         }
-
         if (isParallel) {
-
             // => PARALLEL
-            final DataSource[] datasources = new DataSourceLoader().getDataSources(ids);
+            final DataSource[] dataSources = new DataSourceLoader().getDataSources(ids);
             start = System.currentTimeMillis();
-
             for (int i = 0; i < numIterations; i++) {
-
-                Stream.of(datasources).parallel().forEach(ds -> {
+                Stream.of(dataSources).parallel().forEach(ds -> {
                     try {
                         ds.prepare(workspace);
                     } catch (Exception e) {
@@ -68,16 +60,12 @@ class WorkspaceTest {
             averageExecutionTime = elapsed / numIterations;
             System.out.println("AVG PARALLEL EXECUTION TIME: " + averageExecutionTime + " MS (" +
                                df.format((float) averageExecutionTime) + " SECONDS), " + numIterations + " ITERATIONS");
-
         } else {
-
             // => SEQUENTIAL
-            final DataSource[] datasources = new DataSourceLoader().getDataSources(ids);
+            final DataSource[] dataSources = new DataSourceLoader().getDataSources(ids);
             start = System.currentTimeMillis();
-
             for (int i = 0; i < numIterations; i++) {
-
-                for (DataSource ds : datasources) {
+                for (final DataSource ds : dataSources) {
                     try {
                         ds.prepare(workspace);
                     } catch (Exception e) {
@@ -101,25 +89,20 @@ class WorkspaceTest {
         } catch (IOException e) {
             e.printStackTrace();
         }
-
     }
 
     @ParameterizedTest
     @MethodSource("parallelPoolProvider")
     void processDataSourcesInParallelWithPool(int numThreads) {
-
-        final DataSource[] datasources = new DataSourceLoader().getDataSources(ids);
+        final DataSource[] dataSources = new DataSourceLoader().getDataSources(ids);
         ForkJoinPool customPool = null;
         long start, stop, elapsed;
-
         try {
-
             System.out.println("Creating new custom thread pool with " + numThreads + " threads ... ");
             // init new pool and add task to parse and export each datasource
             customPool = new ForkJoinPool(numThreads);
             start = System.currentTimeMillis();
-            customPool.submit(() -> Stream.of(datasources).parallel().forEach(ds -> {
-
+            customPool.submit(() -> Stream.of(dataSources).parallel().forEach(ds -> {
                 try {
                     ds.prepare(workspace);
                 } catch (Exception e) {
@@ -128,11 +111,9 @@ class WorkspaceTest {
                 ds.parse(workspace);
                 ds.export(workspace);
             })).get();
-
             stop = System.currentTimeMillis();
             elapsed = stop - start;
             System.out.println("=> TIME TAKEN " + (elapsed) + " MS (" + df.format(elapsed / 1000) + " SECONDS)");
-
         } catch (Exception e) {
             e.printStackTrace();
         } finally {
@@ -141,7 +122,6 @@ class WorkspaceTest {
                 customPool.shutdown();
             }
         }
-
     }
 
     static Stream<Arguments> parallelProvider() {
@@ -149,32 +129,7 @@ class WorkspaceTest {
     }
 
     static Stream<Arguments> parallelPoolProvider() {
-        return Stream.of(arguments(2), arguments(4), arguments(6), arguments(Runtime.getRuntime().availableProcessors()));
+        return Stream.of(arguments(2), arguments(4), arguments(6),
+                         arguments(Runtime.getRuntime().availableProcessors()));
     }
-
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-